spark-bq-incremental/requirements.txt
apache-airflow==2.8.1
apache-airflow-providers-google
pyspark==3.4.1
google-cloud-bigquery
google-cloud-storage